Moved sendCategories.js from 'job' to 'steps' folder
