+++ cartridges/int_algolia/cartridge/scripts/algolia/steps/sendCategories.js
